+++ packages/expo-modules-core/android/src/main/java/expo/modules/core/interfaces/ReactNativeHostHandler.java
@@ -1,5 +1,6 @@
 package expo.modules.core.interfaces;
 
+import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import com.facebook.react.bridge.JavaScriptExecutorFactory;
@@ -50,7 +51,7 @@ public interface ReactNativeHostHandler {
    *
    * @return custom DevSupportManagerFactory, or null if not to override
    *
-   * NOTE: This callback is not support in bridgeless mode
+   * NOTE: This callback is not support on bridgeless mode
    */
   @Nullable
   default Object getDevSupportManagerFactory() { return null; }
@@ -58,7 +59,7 @@ public interface ReactNativeHostHandler {
   /**
    * Given chance for modules to override the javascript executor factory.
    *
-   * NOTE: This callback is not support in bridgeless mode
+   * NOTE: This callback is not support on bridgeless mode
    */
   @Nullable
   default JavaScriptExecutorFactory getJavaScriptExecutorFactory() { return null; }
@@ -74,6 +75,13 @@ public interface ReactNativeHostHandler {
    * Callback after react instance creation
    */
   default void onDidCreateReactInstance(boolean useDeveloperSupport, ReactContext reactContext) {}
+
+  /**
+   * Callback when receiving unhandled React Native exceptions
+   *
+   * NOTE: This callback is only available on bridgeless mode
+   */
+  default void onReactInstanceException(boolean useDeveloperSupport, @NonNull Exception exception) {}
 
   //endregion
 }