+++ packages/expo-updates/android/src/main/java/expo/modules/updates/errorrecovery/ErrorRecovery.kt
@@ -10,6 +10,7 @@ import com.facebook.react.bridge.ReactContext
 import com.facebook.react.bridge.ReactMarker
 import com.facebook.react.bridge.ReactMarker.MarkerListener
 import com.facebook.react.bridge.ReactMarkerConstants
+import com.facebook.react.config.ReactFeatureFlags
 import com.facebook.react.devsupport.DisabledDevSupportManager
 import com.facebook.react.devsupport.interfaces.DevSupportManager
 import expo.modules.updates.logging.UpdatesErrorCode
@@ -91,8 +92,13 @@ class ErrorRecovery(
 
   private fun getDevSupportManager(reactContext: ReactContext): DevSupportManager {
     val reactApplication = reactContext.applicationContext as ReactApplication
-    return reactApplication.reactHost?.devSupportManager
-      ?: reactApplication.reactNativeHost.reactInstanceManager.devSupportManager
+    if (ReactFeatureFlags.enableBridgelessArchitecture) {
+      val reactHost = reactApplication.reactHost
+      check(reactHost != null)
+      return reactHost.devSupportManager ?: throw IllegalStateException("Unable to get DevSupportManager from ReactHost")
+    }
+
+    return reactApplication.reactNativeHost.reactInstanceManager.devSupportManager
   }
 
   private fun registerErrorHandler(reactContext: ReactContext) {